Refactor page files
